requirements.txt
click==8.0.3
google-assistant-grpc==0.2.1
google-auth-oauthlib>=0.1.0,<0.3
sounddevice>=0.3.7,<0.4
pathlib2>=2.3.0,<3
pocketsphinx==0.1.15
pypsexec==0.3.0